+++ server/configs/application.properties
@@ -112,10 +112,7 @@ management.endpoint.shutdown.enabled=true
 management.endpoints.enabled-by-default=false
 # allow access via http
 management.endpoints.web.exposure.include=*
-# TODO use a separate port for the actuator.  This requires changes to the LabKeyServer class.
-# The following error is produced:
-#   Exception encountered during context initialization - cancelling refresh attempt: org.springframework.beans.FatalBeanException: ServletWebServerFactory implementation org.labkey.embedded.LabKeyServer$1 cannot be instantiated.
-#   To allow a separate management port to be used, a top-level class or static inner class should be used instead
+# Use a separate port for the actuator
 #management.server.port=8081
 
 # Don't show the Spring banner on startup
